feat: redesign public transfers page to form-first flow
Replaced vehicle-grid-first UX with a lead-capture approach:
hero → inquiry form → optional fleet browsing below.
Clicking a vehicle card pre-selects it in the form and scrolls up.
vehicle_id stays null if no vehicle is selected.
TransferDetailPage now redirects to /transfers.
